+++ CHANGELOG.rst
@@ -6,3 +6,4 @@ Changelog
 
 - Use shortcut CTRL+s,CMD+s or c to save image in slideshow
 - Add help menu, show/hide with h
+- Open image on reddit with o